+++ src/app/auth/facade/auth.facade.ts
@@ -6,6 +6,7 @@ import { ApiService } from '@auth/service/api.service';
 import { StorageService } from '@core/service/storage.service';
 import { TokenService } from '@core/service/token.service';
 import { NavigationRoute } from '@shared/constant/navigation-route.const';
+import { switchMap, tap } from 'rxjs/operators';
 @Injectable({
   providedIn: 'root'
 })
@@ -21,15 +22,32 @@ export class AuthFacade implements IAuthFacade {
 
   login(username: string, password: string): void {
     // Call the API with username and password
-    this.apiService.getUser(username, password).subscribe({
-      next: data => {
-        console.log('API data received:', data);
-        const token = data[0]?.token;
-        this.tokenService.saveToken(token);
-        this.router.navigate([this.navigationRoute.FEATURE.DASHBOARD]);
-        console.log('User logged in');
-      }
-    });
+    this.apiService
+      .getUser(username, password)
+      .pipe(
+        tap(data => {
+          const token = data[0]?.token;
+          const isPrimary = data[0]?.userType === 'Primary';
+
+          this.tokenService.saveToken(token);
+          this.storageService.setItem('is-primary-user', isPrimary);
+        }),
+        switchMap((data: any) => {
+          const role = data[0]?.role;
+          return this.apiService.getUserRoleById(role);
+        })
+      )
+      .subscribe({
+        next: (roleData: any) => {
+          console.log('User role data received:', roleData);
+          this.storageService.setItem('permissions', roleData[0]?.permissions);
+          this.router.navigate([this.navigationRoute.FEATURE.DASHBOARD]);
+          console.log('User logged in');
+        },
+        error: err => {
+          console.error('Login or role fetch failed:', err);
+        }
+      });
   }
 
   logout(): void {