+++ src/app/core/component/sidemenu/sidemenu.component.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import * as MenuItems from '@core/component/sidemenu/json/menu-item.json';
+import { PermissionService } from '@core/service/permission.service';
 
 @Component({
   selector: 'app-sidemenu',
@@ -12,4 +13,6 @@ import * as MenuItems from '@core/component/sidemenu/json/menu-item.json';
 export class SidemenuComponent {
   @Input() isExpanded: boolean = false;
   menuItems = (MenuItems as any).menuItems;
+
+  constructor(public permissionService: PermissionService) {}
 }